# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "APIx - Real-time Airfare Price Index API",
    "version": "1.0.0",
    "description": "Programmatic access to the APIx airfare price index. All endpoints require an X-API-Key header and are rate-limited to 100 req/min/IP."
  },
  "servers": [
    { "url": "https://smart-airfare-price-index.onrender.com/api/v1" }
  ],
  "paths": {
...
    "/admin/coverage": {
      "get": {
        "tags": ["System"],
        "summary": "Scrape coverage and imputation stats per day/portal",
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 14, "maximum": 90 } }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
...
  }
}